Create index.html
The first change that I made was pressing enter after <!DocType html> and adding <html lang="en"> on the second line.  After that, I adjusted lines 4-11 to match Figure 1-35 in the textbook.  Once that was done, I deleted the text between the title tags on line 4, and I typed My First Webpage instead.  Next, I pressed enter twice after my opening body tag to add two new lines, and I typed <p>Welcome to My First Webpage</p>.  To end this step, I pressed the enter key to add a lank line below the paragraph element.  Additionally, I also deleted line 15 because it was blank and not in the figure in the textbook.